requirements.txt
streamlit>=1.28.0
sentence-transformers>=2.2.2
python-docx>=0.8.11
PyPDF2>=3.0.0
numpy>=1.24.0
requests>=2.31.0